Rename files and update healpy version
