+++ zodipy/model_registry.py
@@ -1,10 +1,10 @@
-from zodipy import comps, source_params
+from zodipy import component_params, source_params
 from zodipy.zodiacal_light_model import RRM, Kelsall, model_registry
 
 model_registry.register_model(
     name="dirbe",
     model=Kelsall(
-        comps=comps.DIRBE,
+        comps=component_params.DIRBE,
         spectrum=source_params.SPECTRUM_DIRBE,
         emissivities=source_params.EMISSIVITY_DIRBE,
         albedos=source_params.ALBEDO_DIRBE,
@@ -20,7 +20,7 @@ model_registry.register_model(
 model_registry.register_model(
     name="planck13",
     model=Kelsall(
-        comps=comps.DIRBE,
+        comps=component_params.DIRBE,
         spectrum=source_params.SPECTRUM_PLANCK,
         emissivities=source_params.EMISSIVITY_PLANCK_13,
         T_0=source_params.T_0_DIRBE,
@@ -31,7 +31,7 @@ model_registry.register_model(
 model_registry.register_model(
     name="planck15",
     model=Kelsall(
-        comps=comps.PLANCK,
+        comps=component_params.PLANCK,
         spectrum=source_params.SPECTRUM_PLANCK,
         emissivities=source_params.EMISSIVITY_PLANCK_15,
         T_0=source_params.T_0_DIRBE,
@@ -42,7 +42,7 @@ model_registry.register_model(
 model_registry.register_model(
     name="planck18",
     model=Kelsall(
-        comps=comps.PLANCK,
+        comps=component_params.PLANCK,
         spectrum=source_params.SPECTRUM_PLANCK,
         emissivities=source_params.EMISSIVITY_PLANCK_18,
         T_0=source_params.T_0_DIRBE,
@@ -53,7 +53,7 @@ model_registry.register_model(
 model_registry.register_model(
     name="odegard",
     model=Kelsall(
-        comps=comps.PLANCK,
+        comps=component_params.PLANCK,
         spectrum=source_params.SPECTRUM_PLANCK,
         emissivities=source_params.EMISSIVITY_ODEGARD,
         T_0=source_params.T_0_DIRBE,
@@ -64,7 +64,7 @@ model_registry.register_model(
 model_registry.register_model(
     name="rrm-experimental",
     model=RRM(
-        comps=comps.RRM,
+        comps=component_params.RRM,
         spectrum=source_params.SPECTRUM_IRAS,
         calibration=source_params.CALIBRATION_RRM,
         T_0=source_params.T_0_RRM,
